+++ changelog.md
@@ -69,7 +69,8 @@
 - Add `voice` parameter to `Client.scheduled_event_edit`.
 - Add `location` parameter to `Client.scheduled_event_edit`.
 - Add `status` parameter to `Client.scheduled_event_edit`.
-- `scheduled_event_get_all_guild` endpoint updated.
+- Update `scheduled_event_get_all_guild` updated.
+- Update `scheduled_event_create` endpoint.
 
 #### Bug Fixes